requirements.txt
Flask==2.0.2
pdfminer.six==20201018
supabase==2.9.0
python-dotenv==0.19.2
werkzeug==2.0.2
python-dotenv==0.19.2
